chatbot-backend/pyproject.toml
[project]
name = "chatbot-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.132.0",
    "langchain[google-genai]>=1.2.10",
    "python-dotenv>=1.2.1",
]
